Split code-notification APIs into ICodeNotifications contract
Extract SetCodeNotification, GetCodeNotification, and SetAllCodeNotifications
into a new ICodeNotifications contract, leaving INotifications focused on
read-only event decoding (TryParseNotification, SetGcNotification).

The two responsibilities are materially different: INotifications decodes
exception-record payloads raised by the runtime; ICodeNotifications writes
into an in-process allowlist and lazily allocates the table via AllocVirtual.

Also adds GetCodeNotificationCapacity() to the new contract in preparation
for an upfront capacity check in batch SOSDacImpl.SetCodeNotifications.
