+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Contracts/CoreCLRContracts.cs
@@ -24,6 +24,7 @@ public static class CoreCLRContracts
         registry.Register<IComWrappers>("c1", static t => new ComWrappers_1(t));
         registry.Register<ISHash>("c1", static t => new SHash_1(t));
         registry.Register<INotifications>("c1", static t => new Notifications_1(t));
+        registry.Register<ICodeNotifications>("c1", static t => new CodeNotifications_1(t));
         registry.Register<ISignatureDecoder>("c1", static t => new SignatureDecoder_1(t));
         registry.Register<IBuiltInCOM>("c1", static t => new BuiltInCOM_1(t));
         registry.Register<IConditionalWeakTable>("c1", static t => new ConditionalWeakTable_1(t));
