+++ src/native/managed/cdac/tests/NotificationsTests.cs
@@ -157,343 +157,4 @@ public class NotificationsTests
         Assert.Equal(expectedMethodDesc, catcherEnter.MethodDescAddress.Value);
         Assert.Equal(expectedOffset, catcherEnter.NativeOffset);
     }
-
-    // --- JIT Notification Table Tests ---
-
-    private const uint CLRDATA_METHNOTIFY_NONE = 0;
-    private const uint CLRDATA_METHNOTIFY_GENERATED = 1;
-    private const uint CLRDATA_METHNOTIFY_DISCARDED = 2;
-
-    // JITNotification struct layout for 64-bit LE:
-    //   offset 0: state (ushort, 2 bytes)
-    //   offset 2: 6 bytes padding
-    //   offset 8: clrModule (ulong, 8 bytes)
-    //   offset 16: methodToken (uint, 4 bytes)
-    //   offset 20: 4 bytes padding
-    //   total size: 24 bytes
-    private const int StateOffset = 0;
-    private const int ClrModuleOffset = 8;
-    private const int MethodTokenOffset = 16;
-    private const int EntrySize = 24;
-    private const uint TableCapacity = 10;
-    private const ulong TableAddress = 0x1_0000;
-    private const ulong TablePointerAddress = 0x2_0000;
-
-    private static INotifications CreateContractWithJITTable()
-    {
-        var arch = new MockTarget.Architecture { IsLittleEndian = true, Is64Bit = true };
-        var helpers = new TargetTestHelpers(arch);
-
-        // Build JITNotification type info
-        var typeFields = new Dictionary<string, Target.FieldInfo>
-        {
-            [nameof(Data.JITNotification.State)] = new Target.FieldInfo { Offset = StateOffset },
-            [nameof(Data.JITNotification.ClrModule)] = new Target.FieldInfo { Offset = ClrModuleOffset },
-            [nameof(Data.JITNotification.MethodToken)] = new Target.FieldInfo { Offset = MethodTokenOffset },
-        };
-        var types = new Dictionary<DataType, Target.TypeInfo>
-        {
-            [DataType.JITNotification] = new Target.TypeInfo { Fields = typeFields, Size = EntrySize },
-        };
-
-        // Allocate table memory: (TableCapacity + 1) entries for bookkeeping + actual entries
-        int totalTableSize = EntrySize * ((int)TableCapacity + 1);
-        byte[] tableData = new byte[totalTableSize];
-
-        // Initialize bookkeeping at index 0: length=0, capacity=TableCapacity
-        helpers.Write(tableData.AsSpan(MethodTokenOffset), (uint)0); // length = 0
-        helpers.WritePointer(tableData.AsSpan(ClrModuleOffset), TableCapacity); // capacity
-
-        // Allocate the pointer to the table
-        byte[] tablePointerData = new byte[8];
-        helpers.WritePointer(tablePointerData.AsSpan(), TableAddress);
-
-        var builder = new TestPlaceholderTarget.Builder(arch);
-        builder.MemoryBuilder.AddHeapFragment(new MockMemorySpace.HeapFragment
-        {
-            Address = TableAddress,
-            Data = tableData,
-            Name = "JITNotificationTable"
-        });
-        builder.MemoryBuilder.AddHeapFragment(new MockMemorySpace.HeapFragment
-        {
-            Address = TablePointerAddress,
-            Data = tablePointerData,
-            Name = "JITNotificationTablePointer"
-        });
-
-        builder.AddTypes(types);
-        builder.AddGlobals(
-            (Constants.Globals.JITNotificationTable, TablePointerAddress),
-            (Constants.Globals.JITNotificationTableSize, TableCapacity)
-        );
-        builder.AddContract<INotifications>(version: "c1");
-
-        var target = builder.Build();
-        return target.Contracts.Notifications;
-    }
-
-    [Fact]
-    public void SetCodeNotification_NewEntry_CanBeRead()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module = new(0xAABB_CCDD);
-        uint token = 0x0600_0001;
-
-        contract.SetCodeNotification(module, token, CLRDATA_METHNOTIFY_GENERATED);
-
-        uint result = contract.GetCodeNotification(module, token);
-        Assert.Equal(CLRDATA_METHNOTIFY_GENERATED, result);
-    }
-
-    [Fact]
-    public void GetCodeNotification_NotFound_ReturnsNone()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module = new(0xDEAD);
-        uint token = 0x0600_9999;
-
-        uint result = contract.GetCodeNotification(module, token);
-        Assert.Equal(CLRDATA_METHNOTIFY_NONE, result);
-    }
-
-    [Fact]
-    public void SetCodeNotification_Update_ChangesFlags()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module = new(0x1234);
-        uint token = 0x0600_0001;
-
-        contract.SetCodeNotification(module, token, CLRDATA_METHNOTIFY_GENERATED);
-        Assert.Equal(CLRDATA_METHNOTIFY_GENERATED, contract.GetCodeNotification(module, token));
-
-        contract.SetCodeNotification(module, token, CLRDATA_METHNOTIFY_DISCARDED);
-        Assert.Equal(CLRDATA_METHNOTIFY_DISCARDED, contract.GetCodeNotification(module, token));
-    }
-
-    [Fact]
-    public void SetCodeNotification_ClearEntry_ReturnsNone()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module = new(0x1234);
-        uint token = 0x0600_0001;
-
-        contract.SetCodeNotification(module, token, CLRDATA_METHNOTIFY_GENERATED);
-        Assert.Equal(CLRDATA_METHNOTIFY_GENERATED, contract.GetCodeNotification(module, token));
-
-        contract.SetCodeNotification(module, token, CLRDATA_METHNOTIFY_NONE);
-        Assert.Equal(CLRDATA_METHNOTIFY_NONE, contract.GetCodeNotification(module, token));
-    }
-
-    [Fact]
-    public void SetCodeNotification_MultipleEntries_IndependentlyReadable()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module1 = new(0x1000);
-        TargetPointer module2 = new(0x2000);
-        uint token1 = 0x0600_0001;
-        uint token2 = 0x0600_0002;
-
-        contract.SetCodeNotification(module1, token1, CLRDATA_METHNOTIFY_GENERATED);
-        contract.SetCodeNotification(module2, token2, CLRDATA_METHNOTIFY_DISCARDED);
-
-        Assert.Equal(CLRDATA_METHNOTIFY_GENERATED, contract.GetCodeNotification(module1, token1));
-        Assert.Equal(CLRDATA_METHNOTIFY_DISCARDED, contract.GetCodeNotification(module2, token2));
-    }
-
-    [Fact]
-    public void SetCodeNotification_InvalidFlags_Throws()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module = new(0x1234);
-        uint token = 0x0600_0001;
-
-        Assert.Throws<ArgumentException>(() =>
-            contract.SetCodeNotification(module, token, 0x4));
-    }
-
-    [Fact]
-    public void SetAllCodeNotifications_ClearsAllEntries()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module = new(0x1000);
-        uint token1 = 0x0600_0001;
-        uint token2 = 0x0600_0002;
-
-        contract.SetCodeNotification(module, token1, CLRDATA_METHNOTIFY_GENERATED);
-        contract.SetCodeNotification(module, token2, CLRDATA_METHNOTIFY_GENERATED);
-
-        contract.SetAllCodeNotifications(TargetPointer.Null, CLRDATA_METHNOTIFY_NONE);
-
-        Assert.Equal(CLRDATA_METHNOTIFY_NONE, contract.GetCodeNotification(module, token1));
-        Assert.Equal(CLRDATA_METHNOTIFY_NONE, contract.GetCodeNotification(module, token2));
-    }
-
-    [Fact]
-    public void SetAllCodeNotifications_FilterByModule_ClearsOnlyMatchingEntries()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module1 = new(0x1000);
-        TargetPointer module2 = new(0x2000);
-        uint token = 0x0600_0001;
-
-        contract.SetCodeNotification(module1, token, CLRDATA_METHNOTIFY_GENERATED);
-        contract.SetCodeNotification(module2, token, CLRDATA_METHNOTIFY_GENERATED);
-
-        contract.SetAllCodeNotifications(module1, CLRDATA_METHNOTIFY_NONE);
-
-        Assert.Equal(CLRDATA_METHNOTIFY_NONE, contract.GetCodeNotification(module1, token));
-        Assert.Equal(CLRDATA_METHNOTIFY_GENERATED, contract.GetCodeNotification(module2, token));
-    }
-
-    [Fact]
-    public void SetAllCodeNotifications_UpdateFlags_ChangesAllMatching()
-    {
-        INotifications contract = CreateContractWithJITTable();
-        TargetPointer module = new(0x1000);
-        uint token1 = 0x0600_0001;
-        uint token2 = 0x0600_0002;
-
-        contract.SetCodeNotification(module, token1, CLRDATA_METHNOTIFY_GENERATED);
-        contract.SetCodeNotification(module, token2, CLRDATA_METHNOTIFY_GENERATED);
-
-        contract.SetAllCodeNotifications(TargetPointer.Null, CLRDATA_METHNOTIFY_DISCARDED);
-
-        Assert.Equal(CLRDATA_METHNOTIFY_DISCARDED, contract.GetCodeNotification(module, token1));
-        Assert.Equal(CLRDATA_METHNOTIFY_DISCARDED, contract.GetCodeNotification(module, token2));
-    }
-
-    // --- Null Table / Lazy Allocation Tests ---
-
-    private static INotifications CreateContractWithNullTable(TestPlaceholderTarget.AllocateMemoryDelegate? allocateMemory = null)
-    {
-        var arch = new MockTarget.Architecture { IsLittleEndian = true, Is64Bit = true };
-
-        var typeFields = new Dictionary<string, Target.FieldInfo>
-        {
-            [nameof(Data.JITNotification.State)] = new Target.FieldInfo { Offset = StateOffset },
-            [nameof(Data.JITNotification.ClrModule)] = new Target.FieldInfo { Offset = ClrModuleOffset },
-            [nameof(Data.JITNotification.MethodToken)] = new Target.FieldInfo { Offset = MethodTokenOffset },
-        };
-        var types = new Dictionary<DataType, Target.TypeInfo>
-        {
-            [DataType.JITNotification] = new Target.TypeInfo { Fields = typeFields, Size = EntrySize },
-        };
-
-        // Table pointer is null (all zeros)
-        byte[] tablePointerData = new byte[8];
-
-        var builder = new TestPlaceholderTarget.Builder(arch);
-        builder.MemoryBuilder.AddHeapFragment(new MockMemorySpace.HeapFragment
-        {
-            Address = TablePointerAddress,
-            Data = tablePointerData,
-            Name = "JITNotificationTablePointer"
-        });
-
-        builder.AddTypes(types);
-        builder.AddGlobals(
-            (Constants.Globals.JITNotificationTable, TablePointerAddress),
-            (Constants.Globals.JITNotificationTableSize, TableCapacity)
-        );
-        builder.AddContract<INotifications>(version: "c1");
-
-        if (allocateMemory is not null)
-            builder.UseAllocateMemory(allocateMemory);
-
-        return builder.Build().Contracts.Notifications;
-    }
-
-    [Fact]
-    public void GetCodeNotification_NullTable_ReturnsNone()
-    {
-        INotifications contract = CreateContractWithNullTable();
-        uint result = contract.GetCodeNotification(new TargetPointer(0x1000), 0x0600_0001);
-        Assert.Equal(CLRDATA_METHNOTIFY_NONE, result);
-    }
-
-    [Fact]
-    public void SetAllCodeNotifications_NullTable_NoOp()
-    {
-        INotifications contract = CreateContractWithNullTable();
-        // Should not throw
-        contract.SetAllCodeNotifications(TargetPointer.Null, CLRDATA_METHNOTIFY_NONE);
-    }
-
-    [Fact]
-    public void SetCodeNotification_NullTable_ClearIsNoOp()
-    {
-        INotifications contract = CreateContractWithNullTable();
-        // Clearing on a null table should not throw
-        contract.SetCodeNotification(new TargetPointer(0x1000), 0x0600_0001, CLRDATA_METHNOTIFY_NONE);
-    }
-
-    [Fact]
-    public void SetCodeNotification_NullTable_NoAllocator_Throws()
-    {
-        INotifications contract = CreateContractWithNullTable(allocateMemory: null);
-        Assert.Throws<NotSupportedException>(() =>
-            contract.SetCodeNotification(new TargetPointer(0x1000), 0x0600_0001, CLRDATA_METHNOTIFY_GENERATED));
-    }
-
-    [Fact]
-    public void SetCodeNotification_NullTable_LazyAllocates_ThenWorks()
-    {
-        var arch = new MockTarget.Architecture { IsLittleEndian = true, Is64Bit = true };
-        var helpers = new TargetTestHelpers(arch);
-
-        // Pre-allocate a memory region for the allocator to "return"
-        int totalTableSize = EntrySize * ((int)TableCapacity + 1);
-        byte[] allocatedTableData = new byte[totalTableSize];
-        const ulong AllocatedTableAddress = 0x3_0000;
-
-        var typeFields = new Dictionary<string, Target.FieldInfo>
-        {
-            [nameof(Data.JITNotification.State)] = new Target.FieldInfo { Offset = StateOffset },
-            [nameof(Data.JITNotification.ClrModule)] = new Target.FieldInfo { Offset = ClrModuleOffset },
-            [nameof(Data.JITNotification.MethodToken)] = new Target.FieldInfo { Offset = MethodTokenOffset },
-        };
-        var types = new Dictionary<DataType, Target.TypeInfo>
-        {
-            [DataType.JITNotification] = new Target.TypeInfo { Fields = typeFields, Size = EntrySize },
-        };
-
-        // Table pointer starts as null
-        byte[] tablePointerData = new byte[8];
-
-        var builder = new TestPlaceholderTarget.Builder(arch);
-        builder.MemoryBuilder.AddHeapFragment(new MockMemorySpace.HeapFragment
-        {
-            Address = TablePointerAddress,
-            Data = tablePointerData,
-            Name = "JITNotificationTablePointer"
-        });
-        builder.MemoryBuilder.AddHeapFragment(new MockMemorySpace.HeapFragment
-        {
-            Address = AllocatedTableAddress,
-            Data = allocatedTableData,
-            Name = "AllocatedJITNotificationTable"
-        });
-
-        builder.AddTypes(types);
-        builder.AddGlobals(
-            (Constants.Globals.JITNotificationTable, TablePointerAddress),
-            (Constants.Globals.JITNotificationTableSize, TableCapacity)
-        );
-        builder.AddContract<INotifications>(version: "c1");
-        builder.UseAllocateMemory((size) => new TargetPointer(AllocatedTableAddress));
-
-        var target = builder.Build();
-        INotifications contract = target.Contracts.Notifications;
-
-        TargetPointer module = new(0xAABB_CCDD);
-        uint token = 0x0600_0001;
-
-        // This should lazily allocate the table and set the notification
-        contract.SetCodeNotification(module, token, CLRDATA_METHNOTIFY_GENERATED);
-
-        // Should be able to read it back
-        uint result = contract.GetCodeNotification(module, token);
-        Assert.Equal(CLRDATA_METHNOTIFY_GENERATED, result);
-    }
 }
